WIP: Initial hacking for supplying optional parameter data
This change is to support YCM client generating snippets for function
completions. The idea is to simply provide all of the optional chunks
returned by libclang in a (recursive) structure.

In order to suppport overloads, we remove the "de-duplication" of function
suggestions, and undo @oblitum's change to remove function parameters from
function completion suggestions (as these are required to determine which
overload is being selected).

This allows the user to select the specific overload they want to
complete (and thus generate the correct snippet). Obviously this is contentious
for a number of reaons:

- The original design was clearly to de-duplicate and use the preview window to
  show overloads. Arguably argument hints mean this is no longer the long-term
  design.
- The size of the completion response is increased (and overloads take up
  valuable screen space and contribute towards the maximum number of
  suggestions)
- The user will probably *not* want to use snippet/overload completion in all
  cases, and for certain common system types, there are many overloads which may
  be distracting. As the completion string is still just the function name, this
  doesn't break  muscle memory, but needs to be observed.