Support diagnostics in Tern completer
This requires the 'lint' plugin to be added to the .tern-project file.
See https://github.com/angelozerr/tern-lint for details of usage.

In order to reduce duplication, this imports stuff from the cs_compelter.
FIXME: This should be changed to a diagnostics.py or responses.py

The tests won't pass because I've changed the test file, but that will
be fixed.