Fix TypeScript unicode tests on Windows
TSServer adds a newline at the end of the response message and counts
it as one character (\n) towards the content length. But newlines are
two characters on Windows (\r\n). To take care of that, the
universal_newlines option is set when starting the TSServer
subprocess. This option automatically converts Windows newlines to \n.
However, it also opens the stdin, stdout, and stderr as text streams
instead of binary ones. This does not work properly with unicode
characters on Windows and Python 3. Therefore, we directly increment
the content length if we are on Windows instead of using the
universal_newlines option.