Add subcommands: FixIt, RefactorRename, GoToReferences, GetDoc, GetType
Support for CodeActions via FixIts
Support for RefactorRename
Support for GoToReferences
Add GoTo and GoToDeclaration aliases for GoToDefinition
Implement GetDoc and GetType

GetType is removed from language_server_completer.py since a server is
allowed to return anything upon textDocument/hover request.

GetType and GetDoc are Implemented in java_completer.py with a server
specific modifications to the content of the response to the
textDocument/hover request.