Return diagnostics for all files asynchronously _and_ on parse request.
This changes the diagnostics delivery so that we return the diagnostics
to the client in the same way that the language server does, i.e.
per-file. The client then fans them out or does whatever makes sense for
the client.

In order to be relatively compatible with other clients, we also return
diagnostics on the file-ready-to-parse event, even though they might be
out of date wrt the code. The client is responsible for ignoring these
diagnostics when it handles the asynchronously delivered ones.