New completion handing
We now no longer support completions with newlines.

We now correctly return completions where the entry is a TextEdit and
multiple completions have different start columns.

We do this super inefficiently by attempting to normalise the TextEdits
into insertion_texts with the same start_codepoint. This is necessary
particularly due to the way that eclipse returns import completions for
packages.