Support multi-user installation
Move the config directory inside the workspace, and allow the client to
specify a workspace root directory.  This means that it is possible to
share the ycmd installation across multiple users by moving the mutable
state directories elsewhere.  By default we just move the config dir and
use the old workspace directory.

While we are at it, provide a way to wipe out the re-used workspace.
When use_clean_workspace is False, the workspace can become corrupt, so
we add a command to stop the server, delete the workspace (and
optionally also the config directory), then restart the server.