Support for multi-line completions
This replaces a non-snippet completion containing newlines with a FixIt
which replaces the entered 'trigger' text with the multi-line text.

We use the 'fixit completion' mechanism for this, but include a special
indicator that tells the client to put the cursor position _after_ the
inserted text, as we would for a normal completion.