Return messages from multiple all completers
Previously the message poll (/receive_messages) request only ever asked
a single completer (whichever was associated with the current filetype
in the request).

Incidentally, the YCM (Vim) client only ever created a single poll
request and re-used it for each poll.

This meant that if there were multiple active completers providing async
messages, only messages for the _first_ filetype opened would be
returned. For other clients which might create new request data for each
request, certainly only messages for the active filetype would be
returned.

One option would be to state in the protocol that the messages request
must be targetted at a filetype completer (e.g. by using the
completer_target option), but that is really only for testing. A better
option is to return messages from _all_ completers in the
receive_messages poll, ignoring the request data. In practice, we
actually return messages from the first completer to return from the
poll. As they all use the same timeout, this works fine.

WIP: Unfortunately this makes the code all a bit messy. Perhaps we
should change the way timeouts are handled entirely by removing the
timeout behaviour from the completer classes and just using
concurrent.future.wait() to do that.