Support snippet completions
In general this is terrible, compared to working normal completion, but
some servers (json, yaml) will only ever return snippet completions, and
they will not be convinved to change.

The approach is to put the snippet text in the extra_data and have the
client expand the snippet in-place on selecting the completion item.

As many servers will agressively over-use snippets when the client
claims they support it, we _dont'_ claim we support it, but provide a
mechanism for extra_conf files to override the capabilities that we
declare. This allows users to enable snippets where they make sense,
i.e. for the json LS.