unittest integration improvements
Allow -- to override run_tests.py logic for the unittest command to run.
THis allows for example to run specific path.to.module.ClassTest (which
I use for "run the test under the cursor").

Fix loading of libclang.so on macOS under Python 3.9. For some reason
now with Python 3.9 we need to explicitly call ImportAndCheckCore (i
think this is about loading the "dependencies", but not sure). WIthotu
this change trying to run any test says "starting directory is not
importable", or fails with ImportError etc.

Remove some logger.exception() output and downgrade to logger.warning().
This is rarely actually useful and it really puts a lot of spam in the
tests.