Java: Fix gradle project root detection
Typical "modern" layout from `gradle init`:

- settings.gradle
- app/build.gradle
- ...

If you open app/src/java/.../Foo.java, we incorrectly pick app/ as the
project root.

Now, when searching for the project root, we continue to search the
varioius file names, but we tag each file name with the "kind" and
keep searching for other files of the same *kind* rather than just the
same name.

Meanwhile, we prefer finding the gradle/maven project. The reason for
this is that jdt.ls spams .project files all over the codebase which
makes any project you've previously opened look like an eclipse project.
This is fine if the project root was previously correctly determined
and/or the projects properly imported, but means that we never
self-correct. So prefer finding the non-native project types so that we
pick the correct root in _most_ cases.