Fix jdt.ls configuration quirks
There was a problem with jdt.ls configuration that was bamboozling.

If you put something like this in your config:

```
def Settings(**kwargs):
    if kwargs['language'] == 'java':
        return {
            'ls': { 'settings': {
                'java.project.referencedLibraries': {
                    'include': [
                        '/path/to/some/library.jar'
                    ]
                }
            } }
        }
```

It wouldn't work. Whatever version of this configuration you use it
would just always use the "default" behaviour for
"java.project.referencedLibraries". Debugging jdt.ls with print
statements showed that it was actually receiving a configuration that
didn't contain this.

It turned out this was because it was initially receiving this key, and
correctly reading the config, then subsequently receiving
didChangeConfiguration without this key and reverting its settings to
default.

The workspace/configuration, initializationOptions and
didChangeConfiguration messages in LSP are a total mess, and the
specification is underspecified to say the least. We even have FIXME and
TODOs which link to discussions/bugs on the LSP repo that add more
colour. Anyway it seems that didChangeConfiguration is deprecated,
replaced by workspace/configuration, but as always these "new" things
are optional and it's not clear if it's worth actually doing anyuthing
about that.

But what ycmd does is very simple: whatever is put in the 'ls' entry in
the extra conf goes in initializationOptions and in 'settings' on the
didChangeConfiguration notification, which is always sent after
initialisation (for some reason that amounts to "that's what vscode did
and so was required for some server to work).

Anyway, it turns out that jdt.js has wierd undocumented requirement for
its configurations:

- initializationOptions has keys: "settings", "bundles" and
  "workspaceFolders" (believe it or not)
- the 'settings' entry in the DidChangeConfigurationParams (from the LSP
  spec) contains the 'settings' from above.
- it's not specified (anywhere) what you're supposed to return for
  'workspace/configuration' request (but i can now guess it's the
  'settings' from above), but we don't support that anyway.

So you can see that we were actually sending didChangeConfiguration with
'settings' set to { "bundles": ..., "settings": ... }, and thus it
thought we were _removing all settings_ because they were embedded one
layer too deep.

The solution is to only send 'settings' when sending
didChangeConfiguration.