Support for unsolicited FixIts and document resource operations
This adds support for rename/create/delete of files in response to user
actions and for "workspace edits" to be sent to the client in cases
other than as a result of a YcmCompelter command.

The purpose of this is to support some niche language servers that send
these edits in response to external triggers (in particular this is
used by the Slint language server).

But resource operations are independently useful, creation of classes,
rename of files in response to refactorings, etc. Requires client
support, of course.

We model these unsolicited requests as notifications for the client
because YCM protocol does not have "server to client" requests like LSP
and there's really nothing that a server can do in response to the
rejection anyway. When we get such a request, we put it in the
notifications queue and "ack" the request.